add timezone

+++ database/migrations/2024_08_00_1.5_create_administrations_table.php
@@ -8,15 +8,16 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('administrations', function (Blueprint $table) {
+        if (!Schema::hasTable('administrations')) {
+            Schema::create('administrations', function (Blueprint $table) {
             $table->id();
             $table->string('name')->nullable();
             $table->string('code')->nullable();
             $table->foreignId('manager_id')->nullable()->constrained()->onDelete('set null');
             $table->foreignId('company_id')->nullable()->constrained()->onDelete('cascade');
-
             $table->timestamps();
-        });
+            });
+        }
     }
 
     public function down(): void
